fix: use win32 local cmds for fs op

+++ lua/souvenir.lua
@@ -113,15 +113,27 @@ function M.delete_session(session)
   local session_shada = session..'.shada'
 
   if utils.is_file_exist(session_path..session_file) == true then
-    if os.execute('rm '..session_path..session_file) > 0 then
-      vim.api.nvim_err_writeln('fatal: cannot delete'..session_file..', check your permission!')
-    end
-    if shada ~= false then
-      if os.execute('rm '..SHADA_PATH..session_shada) > 0 then
-        vim.api.nvim_err_writeln('fatal: cannot delete'..session_shada..', check your permission!')
+    if vim.loop.os_uname().version:match('Windows') then
+      if os.execute('del '..session_path..session_file) > 0 then
+        vim.api.nvim_err_writeln('fatal: cannot delete'..session_file..', check your permission!')
       end
+      if shada ~= false then
+        if os.execute('del '..SHADA_PATH..session_shada) > 0 then
+          vim.api.nvim_err_writeln('fatal: cannot delete'..session_shada..', check your permission!')
+        end
+      end
+      vim.api.nvim_echo({{'souvenir: session `'..session..'` deleted', 'Normal'}}, true, {})
+    else
+      if os.execute('rm '..session_path..session_file) > 0 then
+        vim.api.nvim_err_writeln('fatal: cannot delete'..session_file..', check your permission!')
+      end
+      if shada ~= false then
+        if os.execute('rm '..SHADA_PATH..session_shada) > 0 then
+          vim.api.nvim_err_writeln('fatal: cannot delete'..session_shada..', check your permission!')
+        end
+      end
+      vim.api.nvim_echo({{'souvenir: session `'..session..'` deleted', 'Normal'}}, true, {})
     end
-    vim.api.nvim_echo({{'souvenir: session `'..session..'` deleted', 'Normal'}}, true, {})
   else
     vim.api.nvim_err_writeln('fatal: '..session_file..' does not exist')
   end
